Expose Integrations table
