+++ fec.config.js
@@ -1,3 +1,5 @@
+const path = require('path');
+
 module.exports = {
     appUrl: [ '/settings/notifications', 'settings/integrations' ],
     debug: true,
@@ -11,5 +13,11 @@ module.exports = {
      * Add additional webpack plugins
      */
     plugins: [],
-    _unstableHotReload: process.env.HOT === 'true'
+    _unstableHotReload: process.env.HOT === 'true',
+    moduleFederation: {
+        exposes: {
+            './RootApp': path.resolve(__dirname, './src/AppEntry.tsx'),
+            './IntegrationsTable': path.resolve(__dirname, './src/IntegrationsEntry.tsx')
+        }
+    }
 };
