+++ src/IntegrationsEntry.tsx
@@ -9,20 +9,16 @@ import React from 'react';
 import { ClientContextProvider } from 'react-fetching-library';
 import { Provider } from 'react-redux';
 import { BrowserRouter as Router } from 'react-router-dom';
-import * as Redux from 'redux';
 
 import messages from '../locales/data.json';
-import App from './app/App';
+import IntegrationsApp from './app/IntegrationsApp';
+import { AppEntryProps } from './AppEntry';
 import { getNotificationsRegistry } from './store/Store';
 import { getBaseName } from './utils/Basename';
 
-export interface AppEntryProps {
-    logger?: Redux.Middleware;
-}
-
 enableMapSet();
 
-const AppEntry: React.FunctionComponent<AppEntryProps> = (props) => {
+const IntegrationsEntry: React.FunctionComponent<AppEntryProps> = (props) => {
 
     const client = React.useMemo(() => createFetchingClient(getInsights, {
         responseInterceptors: [ validateSchemaResponseInterceptor ]
@@ -38,7 +34,7 @@ const AppEntry: React.FunctionComponent<AppEntryProps> = (props) => {
             <Provider store={ store }>
                 <Router basename={ getBaseName(window.location.pathname) }>
                     <ClientContextProvider client={ client }>
-                        <App />
+                        <IntegrationsApp />
                     </ClientContextProvider>
                 </Router>
             </Provider>
@@ -46,4 +42,4 @@ const AppEntry: React.FunctionComponent<AppEntryProps> = (props) => {
     );
 };
 
-export default AppEntry;
+export default IntegrationsEntry;
